added loop to UserList component
