+++ package.json
@@ -10,6 +10,7 @@
     "axios": "^0.19.2",
     "firebase": "^7.18.0",
     "gh-pages": "^3.1.0",
+    "node-sass": "^4.14.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
