+++ src/index.scss
@@ -1,0 +1,8 @@
+@import './partials/setup';
+@import './partials/variables';
+@import './partials/global';
+@import './partials/body';
+
+/* Normalize */
+
+/* Our Styles */
