+++ src/partials/_global.scss
@@ -1,0 +1,3 @@
+h1 {
+    font-size: 8rem;
+}
